moved stimuli generation for the performance tests to a separate file and reused the data for both, legacy and refactored version
